travis: add script_xvfb_run
Just a simple test run of the program. So far It does not do something
useful on these travis machines.

One may look at the  build logs for OSX. There are some runtime GL warnings
to fix, probably related to the actual problems.

The Linux run on travis is much more broken because the GL version there is
really really old. Probably we don't want to fix that.